zazmol git maintenance

+++ src/python/linear_algebra.py
@@ -116,7 +116,7 @@ def matrix_multiply(a, b):
     except:
         dim_b2 = 1
 
-    c = numpy.zeros((dim_a1, dim_b2), numpy.float)
+    c = numpy.zeros((dim_a1, dim_b2), numpy.float32)
     if(dim_a2 != dim_b1):
         message = 'incompatible matrices'
         error.append(message)
@@ -160,7 +160,7 @@ def find_u(x, y):
 
     '''
 
-    b = numpy.zeros(9, numpy.float)
+    b = numpy.zeros(9, numpy.float32)
     k = 0
     for i in range(3):
         for j in range(3):
@@ -198,11 +198,11 @@ def find_u(x, y):
         urak1 = (1.0 / math.sqrt(abs(uk[1]))) * rak1
 
     urak2 = numpy.cross(urak0, urak1)
-    bk = numpy.zeros((3, 3), numpy.float)
+    bk = numpy.zeros((3, 3), numpy.float32)
     bk[0] = urak0
     bk[1] = urak1
     bk[2] = urak2
-    lu = numpy.zeros(9, numpy.float)
+    lu = numpy.zeros(9, numpy.float32)
     lk = 0
     for j in range(3):
         for i in range(3):
@@ -314,10 +314,10 @@ def dihedral_angle(a1, a2, a3, a4):
 
     '''
 
-    r1 = numpy.zeros(3, numpy.float)
-    r2 = numpy.zeros(3, numpy.float)
-    r3 = numpy.zeros(3, numpy.float)
-    r4 = numpy.zeros(3, numpy.float)
+    r1 = numpy.zeros(3, numpy.float32)
+    r2 = numpy.zeros(3, numpy.float32)
+    r3 = numpy.zeros(3, numpy.float32)
+    r4 = numpy.zeros(3, numpy.float32)
 
     r1 = a1 - a2
     r2 = a3 - a2
